segregating client components
